requirements.txt
aistudio-sdk==0.3.6
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
bce-python-sdk==0.9.45
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
ijson==3.2.3
image-quality==1.2.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
joblib==1.5.1
kiwisolver==1.4.8
label-studio-converter==0.0.59
label-studio-tools==0.0.4
lazy_loader==0.4
libsvm==3.23.0.4
lxml==6.0.0
matplotlib==3.10.3
modelscope==1.29.2
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nvjitlink-cu12==12.9.86
onnxruntime-gpu==1.18.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle-gpu==3.1.1
paddlex==3.2.1
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
proglog==0.1.12
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pybrisque==1.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2025.7.34
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.18.3
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
wcwidth==0.2.13
